Fixed error: incompatible types: possible lossy conversion from double to int                     int wit = Double.parseDouble(sc.nextLine());

+++ Main.java
@@ -17,12 +17,12 @@ class Main {
             switch (cmd) {
                 case "1":
                     System.out.print("\nIevadiet pievienojamo summu: ");
-                    int dep = Double.parseDouble(sc.nextLine());
+                    double dep = Double.parseDouble(sc.nextLine());
                     bankAcount.deposit();
                     break;
                 case "2":
                     System.out.print("\nIevadiet izņemto summu: ");
-                    int wit = Double.parseDouble(sc.nextLine());
+                    double wit = Double.parseDouble(sc.nextLine());
                     bankAcount.withdraw();
                     break;
                 case "3":
